created an MDI page for the forms to load in
the main menu form now is an MDI form so the address input, load file and job forms can now be loaded inside of it creating a significantly cleaner look and feel for using the software. this also fixes issues with closing the incorrect form and the application running infinity in the background.
need to fix issue with the file writer causing a stack overflow error on close but not when saving.