doubled up to 16 cards in the job form
this is to allow the cards to be filled with the proper steps with ease. also need to rename the form objects to represent the associated task, a big task but will streamline code for future changes becoming easier.
need to look into anti file duplication, tested beta version to no avail.