converted file system to use a less error prone path
file system now uses a \\SERVER path instead of a X:\ path. this is due to the difference between computers location of the company files drive not being a consistent X:\ location but all using the correct \\SERVER path underneath

Also moved the MDI parent code for the JobForm into the Load method as opposed to the SetupForm method to make it consistent with the other forms and thus easier to add new forms in the future 

Also project is now being developed on a new 
computer so contains some changes to meta files.